pyproject.toml
[project]
name = "sigan"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "swilcox", email = "steven@wilcoxzone.com" }]
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.7",
    "humanize>=4.11.0",
    "pydantic>=2.9.2",
    "pyyaml>=6.0.2",
    "rich>=13.9.4",
]

[project.scripts]
sigan = "sigan.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "freezegun>=1.5.1",
    "pytest-cov>=6.0.0",
    "pytest>=8.3.3",
]
